requirements.txt
# Core dependencies
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0

# Optional AWS dependencies (for Bedrock integration)
# boto3>=1.28.0
# botocore>=1.31.0

# Utility
python-dateutil>=2.8.2
colorama>=0.4.6  # For colored console output

# Development
pytest>=7.4.0
black>=23.0.0